backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
zhipuai
requests
pydantic
python-dotenv
